no_media in robot/client
